refactor : 관리자 mock 데이터 파일 위치 변경
